p4i integration Integrate change 75423 from 3.0 to main. Description for change 75423:
: Change 75423 by vohi@vohi-hokkaido-nt on 2002/09/27 13:00:37
: 
: 	Silence warnings on 64bit VC++ compiler.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/kernel/qprinter_win.cpp#32 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 75424]

+++ src/kernel/qprinter_win.cpp
@@ -728,9 +728,9 @@ static void setDefaultPrinterA(const QString &printerName, HANDLE *hmode, HANDLE
     }
 
     // Allocate a global HANDLE for a DEVNAMES Structure
-    DWORD lDrvrName = strlen(pinf2->pDriverName) + 1;
-    DWORD lPrntName = strlen(pinf2->pPrinterName) + 1;
-    DWORD lPortName = strlen(pinf2->pPortName) + 1;
+    DWORD lDrvrName = (DWORD)strlen(pinf2->pDriverName) + 1;
+    DWORD lPrntName = (DWORD)strlen(pinf2->pPrinterName) + 1;
+    DWORD lPortName = (DWORD)strlen(pinf2->pPortName) + 1;
     if ( hdevnames ) {
 	GlobalFree( hdevnames );
 	hdevnames = 0;
